+++ Modules/Intent.Modules.Application.Wolverine.FluentValidation/NugetPackages.cs
@@ -18,7 +18,12 @@ namespace Intent.Modules.Application.Wolverine.FluentValidation
             NugetRegistry.Register(FluentValidationPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 8, >= 0) => new PackageVersion("11.9.0"),
+                        ( >= 8, >= 0) => new PackageVersion("12.1.1"),
+                        ( >= 7, >= 0) => new PackageVersion("11.12.0"),
+                        ( >= 6, >= 0) => new PackageVersion("11.12.0"),
+                        ( >= 2, >= 1) => new PackageVersion("11.12.0"),
+                        ( >= 2, >= 0) => new PackageVersion("11.12.0")
+                            .WithNugetDependency("System.Threading.Tasks.Extensions", "4.5.4"),
                         _ => throw new Exception($"Unsupported Framework `{framework.Major}` for NuGet package '{FluentValidationPackageName}'"),
                     }
                 );