+++ Modules/Intent.Modules.AspNetCore.DistributedCaching/NugetPackages.cs
@@ -20,14 +20,14 @@ namespace Intent.Modules.AspNetCore.DistributedCaching
             NugetRegistry.Register(MicrosoftExtensionsCachingAbstractionsPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 10, >= 0) => new PackageVersion("10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Primitives", "10.0.2"),
-                        ( >= 9, >= 0) => new PackageVersion("10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Primitives", "10.0.2"),
-                        ( >= 8, >= 0) => new PackageVersion("10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Primitives", "10.0.2"),
-                        ( >= 2, >= 0) => new PackageVersion("10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Primitives", "10.0.2")
+                        ( >= 10, >= 0) => new PackageVersion("10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Primitives", "10.0.9"),
+                        ( >= 9, >= 0) => new PackageVersion("10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Primitives", "10.0.9"),
+                        ( >= 8, >= 0) => new PackageVersion("10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Primitives", "10.0.9"),
+                        ( >= 2, >= 0) => new PackageVersion("10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Primitives", "10.0.9")
                             .WithNugetDependency("System.Threading.Tasks.Extensions", "4.6.3"),
                         _ => throw new Exception($"Unsupported Framework `{framework.Major}` for NuGet package '{MicrosoftExtensionsCachingAbstractionsPackageName}'"),
                     }
@@ -35,45 +35,45 @@ namespace Intent.Modules.AspNetCore.DistributedCaching
             NugetRegistry.Register(MicrosoftExtensionsCachingMemoryPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 10, >= 0) => new PackageVersion("10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Caching.Abstractions", "10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Options", "10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Primitives", "10.0.2"),
-                        ( >= 9, >= 0) => new PackageVersion("10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Caching.Abstractions", "10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Options", "10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Primitives", "10.0.2"),
-                        ( >= 8, >= 0) => new PackageVersion("10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Caching.Abstractions", "10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Options", "10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Primitives", "10.0.2"),
-                        ( >= 2, >= 0) => new PackageVersion("10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Caching.Abstractions", "10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Options", "10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Primitives", "10.0.2"),
+                        ( >= 10, >= 0) => new PackageVersion("10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Caching.Abstractions", "10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Options", "10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Primitives", "10.0.9"),
+                        ( >= 9, >= 0) => new PackageVersion("10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Caching.Abstractions", "10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Options", "10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Primitives", "10.0.9"),
+                        ( >= 8, >= 0) => new PackageVersion("10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Caching.Abstractions", "10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Options", "10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Primitives", "10.0.9"),
+                        ( >= 2, >= 0) => new PackageVersion("10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Caching.Abstractions", "10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.DependencyInjection.Abstractions", "10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Options", "10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Primitives", "10.0.9"),
                         _ => throw new Exception($"Unsupported Framework `{framework.Major}` for NuGet package '{MicrosoftExtensionsCachingMemoryPackageName}'"),
                     }
                 );
             NugetRegistry.Register(MicrosoftExtensionsCachingStackExchangeRedisPackageName,
                 (framework) => (framework.Major, framework.Minor) switch
                     {
-                        ( >= 10, >= 0) => new PackageVersion("10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Caching.Abstractions", "10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Options", "10.0.2")
+                        ( >= 10, >= 0) => new PackageVersion("10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Caching.Abstractions", "10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Options", "10.0.9")
                             .WithNugetDependency("StackExchange.Redis", "2.7.27"),
-                        ( >= 2, >= 0) => new PackageVersion("10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Caching.Abstractions", "10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "10.0.2")
-                            .WithNugetDependency("Microsoft.Extensions.Options", "10.0.2")
+                        ( >= 2, >= 0) => new PackageVersion("10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Caching.Abstractions", "10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Logging.Abstractions", "10.0.9")
+                            .WithNugetDependency("Microsoft.Extensions.Options", "10.0.9")
                             .WithNugetDependency("StackExchange.Redis", "2.7.27"),
                         _ => throw new Exception($"Unsupported Framework `{framework.Major}` for NuGet package '{MicrosoftExtensionsCachingStackExchangeRedisPackageName}'"),
                     }